requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
bcrypt
pyjwt
langchain
langchain-core
langchain_community
langchain-classic
langchain-groq
pydantic
pydantic[email]
dotenv
scikit-learn
numpy
regex
flask
twilio
python-multipart
motor
gunicorn
flask-cors